+++ src/onegov.notice/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Adds an 'accepted' state.
+  [msom]
+
 0.0.1 (2017-07-14)
 ~~~~~~~~~~~~~~~~~~